tiktok-reply/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
sqlite3
openai==1.3.0
flask==2.3.3
beautifulsoup4==4.12.2
selenium==4.15.0
webdriver-manager==4.0.1
pandas==2.1.1
logging 